23995 - extract common method
