+++ src/NHSD.GPIT.BuyingCatalogue.UI.Components/Views/Shared/TagHelpers/Table/TableRowHeaderCellTagHelper.cs
@@ -6,15 +6,13 @@ using NHSD.GPIT.BuyingCatalogue.UI.Components.TagHelpers;
 
 namespace NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.TagHelpers.Table
 {
-    [HtmlTargetElement(TagHelperName, ParentTag = TableRowContainerTagHelper.TagHelperName)]
+    [HtmlTargetElement(TableRowHeaderCellTagHelper.TagHelperName, ParentTag = TableRowContainerTagHelper.TagHelperName)]
     public sealed class TableRowHeaderCellTagHelper : TagHelper
     {
         private const string TagHelperName = "nhs-table-row-header-cell";
 
         private const string RowHeaderRole = "rowheader";
         private const string RowHeaderScope = "row";
-
-        private const string HeadingClass = "nhsuk-table-responsive__heading";
 
         public override async Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
         {
@@ -24,31 +22,13 @@ namespace NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.TagHelpers.Table
             output.Attributes.Add(new TagHelperAttribute(TagHelperConstants.Role, RowHeaderRole));
             output.Attributes.Add(new TagHelperAttribute(TagHelperConstants.Scope, RowHeaderScope));
 
-            var heading = GetHeadingBuilder(context);
+            var heading = CellHeadingBuilder.GetHeadingBuilder(context);
 
             var children = await output.GetChildContentAsync();
 
             output.Content
                 .AppendHtml(heading)
                 .AppendHtml(children);
-        }
-
-        private static TagBuilder GetHeadingBuilder(TagHelperContext context)
-        {
-            if (!context.Items.TryGetValue(TagHelperConstants.CellColumnContextName, out object columnNames))
-                return null;
-
-            var columnNamesConverted = (Queue<TagHelperContent>)columnNames;
-
-            if (columnNamesConverted.Count == 0)
-                return null;
-
-            var builder = new TagBuilder(TagHelperConstants.Span);
-            builder.AddCssClass(HeadingClass);
-
-            builder.InnerHtml.AppendHtml(columnNamesConverted.Dequeue());
-
-            return builder;
         }
     }
 }
